+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.7.0] - not released
+
+> Big thanks to @cugu and @Code0x58
+
 ### Added
 - integration tests on windows, ubuntu, and mac for github
 ### Changed