+++ CHANGELOG.md
@@ -3,7 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [1.7.0] - not released
+## [1.x.x] - Unreleased
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [1.7.0] - 9/2/21
 
 > Big thanks to @cugu and @Code0x58